refactor: send CreatableSelect as a prop
